+++ src/TextLabel.js
@@ -1,11 +1,12 @@
 import React from 'react';
 import classNames from 'classnames';
+import prefixClass from './utils/prefixClass';
 import rowComp from './mixins/rowComp';
 import './styles/TextLabel.scss';
 
 import RowCompBody from './RowCompBody';
 
-export const COMPONENT_NAME = 'ic-text-label';
+export const COMPONENT_NAME = prefixClass('text-label');
 
 function TextLabel({ className, children, ...otherProps }) {
     const rootClassName = classNames(className, COMPONENT_NAME);