+++ src/mixins/rowComp.js
@@ -36,7 +36,8 @@ import '../styles/RowComp.scss';
 
 import getComponentName from '../utils/getComponentName';
 import icBEM from '../utils/icBEM';
-import icState from '../utils/icState';
+import prefixClass from '../utils/prefixClass';
+import prefixState from '../utils/prefixState';
 import { statusPropTypes } from './withStatus';
 
 import Icon from '../Icon';
@@ -44,16 +45,16 @@ import Text from '../Text';
 
 import { STATUS_CODE } from '../StatusIcon';
 
-export const COMPONENT_NAME = 'ic-row-comp';
+export const COMPONENT_NAME = prefixClass('row-comp');
 const ROOT_BEM = icBEM(COMPONENT_NAME);
 export const ROW_COMP_BODY = ROOT_BEM.element('body');
 
 // State class names
-const CLASS_ACTIVE = icState('active');
-const CLASS_HIGHLIGHT = icState('highlight');
-const CLASS_ERROR = icState('error');
-const CLASS_DISABLED = icState('disabled');
-const CLASS_UNTOUCHABLE = icState('untouchable');
+const CLASS_ACTIVE = prefixState('active');
+const CLASS_HIGHLIGHT = prefixState('highlight');
+const CLASS_ERROR = prefixState('error');
+const CLASS_DISABLED = prefixState('disabled');
+const CLASS_UNTOUCHABLE = prefixState('untouchable');
 
 // Alignments
 const LEFT = 'left';