+++ CHANGELOG.md
@@ -9,12 +9,15 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 - Install `prop-types` package in `dependencies` as official recommended.
 - Add new search and CRM icons.
 - Add new `<SearchInput>`.
+- Prefix component class names with `prefixClass()` helper.
 
 ### Changed
 - Upgrade `react` & `react-dom` to v15.5.
 - Migrate `PropTypes` to `prop-types` package instead of import it from main `React` object.
 - Migrate `React.createClass` to `createReactClass` from `create-react-class` package.
 - Upgrade `enzyme` to 2.8.2 to support React v15.5, also install `react-test-renderer`.
+- `icState()` is renamed to `prefixState()` and add prefix with `prefixClass()` helper.
+- ClassName and component prefixes are now `gyp-`
 
 ### Removed
 - Remove deprecated `react-addons-test-utils` package.